UI updates for setting rh value

+++ src/Structs.h
@@ -26,7 +26,7 @@ struct Message{
     char string[64];
     double temp = 0.0;
     double rh = 0.0;
-    int32_t target_rh;
+    uint8_t target_rh;
 };
